+++ Src/Roflcopter.Plugin.Tests/MismatchedFileNames/MismatchedFileNameHighlightingQuickFixTest.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using JetBrains.ProjectModel;
+using JetBrains.ReSharper.Feature.Services.Daemon;
 using JetBrains.ReSharper.Feature.Services.QuickFixes;
 using JetBrains.ReSharper.FeaturesTestFramework.Intentions;
 using JetBrains.ReSharper.TestFramework;
@@ -15,9 +16,9 @@ namespace Roflcopter.Plugin.Tests.MismatchedFileNames
     {
         private TestQuickFix _quickFix;
 
-        protected override IQuickFix CreateQuickFix(IProject project, ITextControl textControl)
+        protected override IQuickFix CreateQuickFix(IProject project, ITextControl textControl, out IHighlighting highlighting)
         {
-            _quickFix = (TestQuickFix) base.CreateQuickFix(project, textControl);
+            _quickFix = (TestQuickFix) base.CreateQuickFix(project, textControl, out highlighting);
             return _quickFix;
         }