+++ Src/Roflcopter.Plugin.Tests/UpdateAspDesignerFiles/UpdateAspDesignerFileQuickFixTest.cs
@@ -29,7 +29,7 @@ namespace Roflcopter.Plugin.Tests.UpdateAspDesignerFiles
         {
             base.DoTest(lifetime, testProject);
 
-            var designerFile = testProject.GetPsiSourceFileInProject(_aspCodeBehindTestHelper.DesignerFileName.ToFileSystemPath());
+            var designerFile = testProject.GetPsiSourceFileInProject(_aspCodeBehindTestHelper.DesignerFileName.ToFileSystemPath().ToVirtualFileSystemPath());
 
             ExecuteWithGold(_aspCodeBehindTestHelper.DesignerFileName, writer => { writer.Write(designerFile.Document.GetText()); });
         }