+++ Src/Roflcopter.Plugin.Tests/test/data/UpdateAspDesignerFiles/ManualTest.aspx.cs
@@ -1,0 +1,8 @@
+﻿using System.Web.UI;
+
+namespace Roflcopter.Sample.UpdateAspDesignerFiles
+{
+    public partial class ManualTest : Page
+    {
+    }
+}