+++ Src/Roflcopter.Plugin/UnitTesting/UnitTestRunFileAction.cs
@@ -5,9 +5,9 @@ using JetBrains.Application.UI.ActionSystem.ActionsRevised.Menu;
 using JetBrains.ProjectModel;
 using JetBrains.ProjectModel.DataContext;
 using JetBrains.ReSharper.Feature.Services.Menu;
-using JetBrains.ReSharper.UnitTestFramework;
 using JetBrains.ReSharper.UnitTestFramework.Actions;
 using JetBrains.ReSharper.UnitTestFramework.Criteria;
+using JetBrains.ReSharper.UnitTestFramework.Execution.Launch;
 using JetBrains.ReSharper.UnitTestFramework.Resources;
 
 namespace Roflcopter.Plugin.UnitTesting