+++ History.md
@@ -1,5 +1,6 @@
 ### 1.14.0 ###
 - Drop the "Copy Fully-qualified name/ Source browser URI to Clipboard" providers feature because it's now part of R# 2021.3
+- ReSharper and Rider 2021.3 support
 
 ### 1.13.0 ###
 - "Update ASP.NET Designer File" context action + reference error quick fix